# Drop hard dependency on Acquisition and speed up removeAllProxies