--- a/setup.py
+++ b/setup.py
@@ -47,7 +47,6 @@ setup(
     tests_require=TESTS_REQUIRE,
     install_requires=[
         'setuptools',
-        'Acquisition',
         'BTrees',
         'isodate',
         'nti.schema',