@jamadden 2017-09-28
There's nothing wrong with Acquisition, it can be extremely helpful. There was just no reason for this project to actually have a dependency on it---it doesn't need to.

I don't think you can monkey-patch Acquisition in because it's a C base class (typically). But again, this project wasn't doing anything like that and this change should have no impact on deployed applications---if they were already using Acquisition, they still can with no changes.